feat: add support for custom default chat configuration
