+++ composables/store.ts
@@ -1,0 +1,7 @@
+import { useStorage } from '@vueuse/core'
+import { DEFAULT_ATTACHED_MESSAGES_COUNT } from '~/config'
+
+export const chatDefaultSettings = useStorage('chat-default-settings', {
+  model: '',
+  attachedMessagesCount: DEFAULT_ATTACHED_MESSAGES_COUNT,
+})
